+++ packages/hid_api/example/linux/flutter/generated_plugins.cmake
@@ -7,7 +7,7 @@ list(APPEND FLUTTER_PLUGIN_LIST
   file_selector_linux
   flutter_secure_storage_linux
   gtk
-  hid_api
+  hid_api_linux
   url_launcher_linux
 )